requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
pre-commit==3.5.0

# Debugging
ipdb==0.13.11
django-debug-toolbar==4.2.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# API testing
requests==2.31.0

# Environment management
python-dotenv==1.0.0
